move events type to documentation
